Add files via upload
This project is called Insta-Liker. It’s a script with the sole purpose of quickly and efficiently liking a fixed number of posts, of a mentioned profile, all without the user interacting during the process. It uses NodeJS, with the puppeteer module to automate the process. The process consist of opening a browser window, redirecting to the Instagram page, logging in with preset credentials(provided in a different file), looking up the aforementioned profile, and then liking the given number of posts by the profile.